Move shell to own component
